Explicación sobre Dakar

+++ Clase 11/Dakar/src/com/company/Auto.java
@@ -2,8 +2,8 @@ package com.company;
 
 public class Auto extends Vehiculo {
 
-    public Auto(Integer velocidad, Integer aceleracion, Integer anguloDeGiro, String patente) {
-        super(velocidad, aceleracion, anguloDeGiro, patente, 1000, 4);
+    public Auto(Double velocidad, Double aceleracion, Double anguloDeGiro, String patente) {
+        super(velocidad, aceleracion, anguloDeGiro, patente, 1000.0, 4);
     }
