+++ Clase 11/Dakar/src/com/company/Moto.java
@@ -3,8 +3,8 @@ package com.company;
 public class Moto extends Vehiculo {
 
 
-    public Moto(Integer velocidad, Integer aceleracion, Integer anguloDeGiro, String patente) {
-        super(velocidad, aceleracion, anguloDeGiro, patente, 300, 2);
+    public Moto(Double velocidad, Double aceleracion, Double anguloDeGiro, String patente) {
+        super(velocidad, aceleracion, anguloDeGiro, patente, 300.0, 2);
 
     }
